migrated to mariadb

+++ back/app/src/models/elective.py
@@ -1,18 +1,18 @@
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from src.database import Base
-from sqlalchemy import BigInteger, ForeignKey
+from sqlalchemy import BigInteger, Integer, ForeignKey
 
 
 class ElectiveModel(Base):
     __tablename__ = "elective"
 
     student_id: Mapped[int] = mapped_column(
-        BigInteger,
+        Integer,
         ForeignKey("user.id", ondelete="CASCADE"),
         primary_key=True
     )
     discipline_id: Mapped[int] = mapped_column(
-        BigInteger,
+        Integer,
         ForeignKey("discipline.id", ondelete="CASCADE"),
         primary_key=True
     )
